+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fj-bom version set to 1.4.5
 - fj-core version set to 8.3.8
 - fj-doc version set to 3.0.9
+- Removed javax.ejb.Stateles annotation from default rest generator
 
 ### Removed